refactor: split all files in smaller parts for better overview like the moonraker docs
